refactor: move repo date and time to schedule
